requirements.txt
asgiref==3.8.1
Django==5.0.5
django-nested-admin==4.0.2
pillow==10.3.0
python-monkey-business==1.0.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
